Renamed api package and reenabled tests in ContigAliasControllerTest
